Reformat to setDisplay to use secondNum for text
This fixes a bug that existed when using sequenced
operator presses. Before, b/c the operatorActive
value stayed at true, you could not enter more
than one digit into the display once you pressed
the operator a second time. Now, on the second
(or otherwise sequential) operator press, the
secondNum variable is set to an empty string after
it is used in the calculation. That way, when we
want to add to the display later, we can use
secondNum as an empty dummy variable to carry the
new numbers being entered by the user, and then
set the display to secondNum. This avoids the
problem of having to clear display.textContent,
instead we first clear the secondNum variable w/in
the operatorPress function, then add our new
values to it in the setDisplay function, and
finally set display.textContent to secondNum!